Combine test_ert_run_context test_run_context
